requirements.txt
gradio==4.41.0
openai-whisper==20231117
torch==2.2.2
gTTS==2.5.1
groq==0.9.0
